Extract visible_from_above_conditions from group based readables
Since the condition only make sense when fetching people, it shouldn't
be part of the `GroupBasedFetchables` class. For example, in
hitobito/hitobito_dsj#34, `GroupBasedFetchables` is used to fetch events.
